Use maximum of 10 sockets to make requests
Seeing some issues when getting a large set of data where node can't get
DNS queries fast enough. This limits us to 10 concurrent requests, which
are held open with keep-alive. This also adds some debugging so you can
see when requests are successful rather than just when they are added to
the queue.